# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: PhishNet API
  description: |
    API REST de PhishNet, plateforme B2B de simulation de phishing.
    **Authentification :** cookie HttpOnly `access_token_cookie` positionné par le serveur
    via Flask-JWT-Extended 4.7.1. Le frontend envoie `credentials: 'include'` à chaque requête.
  version: "1.0.0"
  contact:
    name: Diogo Miguel Gomes Lopes
    email: diogomiguelgl@gmail.com
servers:
  - url: http://localhost:5000
    description: Développement local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
          format: date-time
    Invitation:
      type: object
      properties:
        id:
          type: integer
        invitation_code:
          type: string
        tenant_id:
          type: integer
        is_used:
          type: boolean
        email:
          type: string
          format: email
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
          nullable: true
        user_email:
          type: string
        user_name:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        details:
          type: string
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
  responses:
    Unauthorized:
      description: Token JWT absent, expiré ou révoqué
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            msg: "Missing JWT in cookies or headers"
    Forbidden:
      description: Permission insuffisante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Permission denied"
    NotFound:
      description: Ressource non trouvée (ou appartient à un autre tenant)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Campaign not found"
...
security:
  - cookieAuth: []
paths:
...
  /api/tenants:
    get:
      tags: [Tenants]
      summary: Liste tous les tenants (super-admin uniquement)
      responses:
        "200":
          description: Liste des tenants
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
        "403":
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Tenants]
      summary: Créer un tenant (super-admin uniquement)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                invitation_expires_days:
                  type: integer
                  default: 7
      responses:
        "201":
          description: Tenant créé avec code d'invitation initial
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant:
                    $ref: '#/components/schemas/Tenant'
                  invitation:
                    $ref: '#/components/schemas/Invitation'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/tenants/{id}:
    get:
      tags: [Tenants]
      summary: Détail d'un tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tenants]
      summary: Renommer un tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Tenant modifié
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  tenant:
                    $ref: '#/components/schemas/Tenant'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tenants]
      summary: Supprimer un tenant (bloqué si des utilisateurs existent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant supprimé
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          description: Impossible de supprimer un tenant avec des utilisateurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Cannot delete tenant with associated users"
        "403":
          $ref: '#/components/responses/Forbidden'

  # ── INVITATIONS ─────────────────────────────────────────────────────────────
  /api/tenant-invitations:
    post:
      tags: [Invitations]
      summary: Générer un code d'invitation
      description: |
        Permission requise, `manage_team`.
        Si `email` est fourni, un email SMTP est envoyé avec le code d'invitation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id]
              properties:
                tenant_id:
                  type: integer
                expires_days:
                  type: integer
                  default: 7
                email:
                  type: string
                  format: email
                  nullable: true
      responses:
        "201":
          description: Invitation créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  email_sent:
                    type: boolean
                  invitation:
                    $ref: '#/components/schemas/Invitation'
        "400":
          description: Champ requis manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/tenant-invitations/validate:
    post:
      tags: [Invitations]
      summary: Valider un code d'invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invitation_code]
              properties:
                invitation_code:
                  type: string
      responses:
        "200":
          description: Code valide
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  invitation:
                    $ref: '#/components/schemas/Invitation'
        "400":
          description: Code invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tenant-invitations/{code}:
    get:
      tags: [Invitations]
      summary: Récupérer une invitation par son code
      security: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/tenant-invitations/tenant/{tenant_id}:
    get:
      tags: [Invitations]
      summary: Liste des invitations d'un tenant
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liste des invitations du tenant
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invitation'
        "401":
          $ref: '#/components/responses/Unauthorized'

  # ── AUDIT LOGS ──────────────────────────────────────────────────────────────
  /api/audit-logs:
    get:
      tags: [Audit Logs]
      summary: Journal d'audit paginé (admin uniquement)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: user_id
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Logs paginés
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
                  total_pages:
                    type: integer
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/audit-logs/export:
    get:
      tags: [Audit Logs]
      summary: Exporter les logs en CSV
      responses:
        "200":
          description: Fichier CSV
          headers:
            Content-Disposition:
              schema:
                type: string
                example: "attachment; filename=audit_logs.csv"
          content:
            text/csv:
              schema:
                type: string
        "403":
          $ref: '#/components/responses/Forbidden'

  # ── TRACKING (public) ───────────────────────────────────────────────────────
...